Core: Updated to 10.0.2
Port From (https://github.com/TrinityCore/TrinityCore/commit/e98e1283ea0034baf6be9aa2ffb386eb5582801b)

+++ Source/Game/BattleGrounds/BattleGroundManager.cs
@@ -860,12 +860,12 @@ namespace Game.BattleGrounds
 
         public ushort GetMinPlayersPerTeam()
         {
-            return BattlemasterEntry.MinPlayers;
+            return (ushort)BattlemasterEntry.MinPlayers;
         }
 
         public ushort GetMaxPlayersPerTeam()
         {
-            return BattlemasterEntry.MaxPlayers;
+            return (ushort)BattlemasterEntry.MaxPlayers;
         }
 
         public byte GetMinLevel()
